ajoute fonction de errorhandler et   documentation sur readme.md
