+++ README.md
@@ -1,0 +1,27 @@
+## run the app
+
+> this app is under continuous development, so this instructions may change later
+
+- First clone the project
+
+```bash
+git clone https://github.com/khalidhagane/PharmGarde_mobile
+```
+
+- Inside the project you have two layers, client and server, locate to each directory and install all needed dependencies
+
+  ```bash
+  cd frontend
+  cd web
+  npm install
+  npm start
+  ```
+
+  ```bash
+  cd server
+  npm install
+  npm run server
+  ```
+
+
+- the app will be running on http://localhost:3000
